refactor. move Article, Comment package
